+++ test/combo/template/ceex_engine/debug_annotation_test/setup.exs
@@ -1,3 +1,0 @@
-Combo.Env.put_env(:template, :ceex_debug_annotations, true)
-Code.require_file("components.exs", __DIR__)
-Combo.Env.put_env(:template, :ceex_debug_annotations, false)